agents/shared/pyproject.toml
[project]
name = "council-shared"
version = "0.1.0"
description = "Shared library for The Council's A2A agents — Agent Card factory, FHIR hook, MCP client, audit log."
requires-python = ">=3.11,<3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.10,<1.0.0",
    "google-adk>=1.25.0",
    "google-genai>=1.0.0",
    "mcp>=1.9.0",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
    "starlette>=0.46.0",
    "uvicorn[standard]>=0.34.0",
    "supabase>=2.10.0",
    "sentry-sdk[starlette]>=2.18.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/council_shared"]
